+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 ### Added
+- Added getVarFarkasCoef() (untested)
 ### Fixed
 - all fundamental callbacks now raise an error if not implemented
+- getTermsQuadratic() now correctly returns all linear terms
 ### Changed
 - changed default value of enablepricing flag to True
 ### Removed